+++ apps/client/src/app/services/import-transactions.service.ts
@@ -37,9 +37,9 @@ export class ImportTransactionsService {
       skipEmptyLines: true
     }).data;
 
-    const orders: CreateOrderDto[] = [];
+    const activities: CreateOrderDto[] = [];
     for (const [index, item] of content.entries()) {
-      orders.push({
+      activities.push({
         accountId: this.parseAccount({ item, userAccounts }),
         currency: this.parseCurrency({ content, index, item }),
         dataSource: this.parseDataSource({ item }),
@@ -52,13 +52,13 @@ export class ImportTransactionsService {
       });
     }
 
-    await this.importJson({ content: orders });
+    await this.importJson({ content: activities });
   }
 
   public importJson({ content }: { content: CreateOrderDto[] }): Promise<void> {
     return new Promise((resolve, reject) => {
       this.postImport({
-        orders: content
+        activities: content
       })
         .pipe(
           catchError((error) => {
@@ -121,7 +121,10 @@ export class ImportTransactionsService {
       }
     }
 
-    throw { message: `orders.${index}.currency is not valid`, orders: content };
+    throw {
+      activities: content,
+      message: `activities.${index}.currency is not valid`
+    };
   }
 
   private parseDataSource({ item }: { item: any }) {
@@ -164,7 +167,10 @@ export class ImportTransactionsService {
       }
     }
 
-    throw { message: `orders.${index}.date is not valid`, orders: content };
+    throw {
+      activities: content,
+      message: `activities.${index}.date is not valid`
+    };
   }
 
   private parseFee({
@@ -184,7 +190,10 @@ export class ImportTransactionsService {
       }
     }
 
-    throw { message: `orders.${index}.fee is not valid`, orders: content };
+    throw {
+      activities: content,
+      message: `activities.${index}.fee is not valid`
+    };
   }
 
   private parseQuantity({
@@ -204,7 +213,10 @@ export class ImportTransactionsService {
       }
     }
 
-    throw { message: `orders.${index}.quantity is not valid`, orders: content };
+    throw {
+      activities: content,
+      message: `activities.${index}.quantity is not valid`
+    };
   }
 
   private parseSymbol({
@@ -224,7 +236,10 @@ export class ImportTransactionsService {
       }
     }
 
-    throw { message: `orders.${index}.symbol is not valid`, orders: content };
+    throw {
+      activities: content,
+      message: `activities.${index}.symbol is not valid`
+    };
   }
 
   private parseType({
@@ -255,7 +270,10 @@ export class ImportTransactionsService {
       }
     }
 
-    throw { message: `orders.${index}.type is not valid`, orders: content };
+    throw {
+      activities: content,
+      message: `activities.${index}.type is not valid`
+    };
   }
 
   private parseUnitPrice({
@@ -276,12 +294,12 @@ export class ImportTransactionsService {
     }
 
     throw {
-      message: `orders.${index}.unitPrice is not valid`,
-      orders: content
+      activities: content,
+      message: `activities.${index}.unitPrice is not valid`
     };
   }
 
-  private postImport(aImportData: { orders: CreateOrderDto[] }) {
+  private postImport(aImportData: { activities: CreateOrderDto[] }) {
     return this.http.post<void>('/api/v1/import', aImportData);
   }
 }